Implement a timer that persists through save/load
Added a timer to the main game class, that starts when the first move
is made. Refactored saving and loading to save the state of MineGame,
not Board, so that the timer can be saved. Refactored loading logic to
use a factory method so that loading logic is in the game class, not
the launch script.